+++ src/main/scala/draco/drake/Presence.scala
@@ -2,7 +2,9 @@ package draco.drake
 
 import draco._
 
-sealed trait Presence[T] extends DracoType
+sealed trait Presence[T] extends DracoType {
+  def fold[R](absent: R, present: T => R): R
+}
 
 object Presence extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("Presence", _namePackage = Seq ("draco", "drake")))
@@ -12,6 +14,7 @@ object Presence extends App with DracoType {
 
 trait Present[T] extends Presence[T] {
   val value: T
+  def fold[R](absent: R, present: T => R): R = present(value)
 }
 
 object Present extends App with DracoType {
@@ -33,7 +36,9 @@ object Present extends App with DracoType {
 
 }
 
-trait Absent[T] extends Presence[T]
+trait Absent[T] extends Presence[T] {
+  def fold[R](absent: R, present: T => R): R = absent
+}
 
 object Absent extends App with DracoType {
   override lazy val typeDefinition: TypeDefinition = TypeLoader.loadType(TypeName ("Absent", _namePackage = Seq ("draco", "drake")))